separate hal_disable_irq so the code can be reused. (This allows it to be statically linked in each module that needs it, reducing call overhead.)
